Remove geography option
Specifying geography/geometry and SRID in a setting doesn't work well.
The schema migration has to pick one, and it picked the geography used
for testing.  Geometry in SRID 4326 (WGS84) is a good option for most
GTFS work, and we can't pick a better local projection and still
support all the world's GTFS feeds.